create a class for the main grid
