+++ biogrid/libs/biogrid-simulator/src/lib/biogrid-state/index.ts
@@ -1,1 +1,2 @@
 export * from './biogrid-state-class';
+export * from './main-grid-item';
